Add files via upload
BMI Is A Number Calculated From An Individual’s Weight And Height. To Find Out BMI We Will Take Input From The User (Both Height And Weight) Which Will Be Stored In Height And Weight Variables For Further Calculation. The Calculation Process Is Simple, We Will Simply Divide Weight In Kilograms By The Square Of The Height. Now As Per The BMI Calculated, It Will Execute The Respective If-Else Statement. We Are Also Checking If The User Is Pressing Submit Button Without Entering The Inputs, In That Case, We Are Printing Provide Height Or Provide Weight. Using HTML We Are Giving Desired Structure, An Option For The Input, And Submit Button. With The Help Of CSS, We Are Beautifying Our Structure By Giving Colors And Desired Font, Etc. In The JavaScript Section, We Are Processing The Taken Input And After Calculating, The Respective Output Is Printed.